Initial commit
the flask server files come with an additional test_emotion.py and test_emotion_all.py which allows the testing of inference without needing to host the app.py Flask server. This version appends whitespace to the bottom of processed images and adds the emotion that the model detects in text.